Backend/requirements.txt
flask
python-dotenv
google-generativeai
sentence-transformers
faiss-cpu
requests
python-telegram-bot
flask-cors
bleach
